ML/requirements-ml.txt
fastapi>=0.110.0
uvicorn>=0.27.1
python-multipart>=0.0.9
PyPDF2>=3.0.1
python-docx>=1.1.0
spacy>=3.7.4
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
transformers>=4.38.2
torch>=2.2.1
scikit-learn>=1.4.1.post1
xgboost>=2.0.3
sentence-transformers>=2.5.1
faiss-cpu>=1.8.0
numpy>=1.26.4
pandas>=2.2.1
google-generativeai>=0.4.1
python-dotenv>=1.0.1
sentencepiece==0.1.99
protobuf==4.25.3
